pyproject.toml
[project]
name = "alertiq"
version = "0.1.0"
description = "Intelligent AI-powered alert email processing and routing system"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.3.0",
    "google-auth>=2.23.0",
    "google-auth-oauthlib>=1.0.0",
    "google-auth-httplib2>=0.1.1",
    "google-api-python-client>=2.100.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.4.0",
    "pydantic-settings>=2.0.0",
    "loguru>=0.7.0",
    "tenacity>=8.2.0",
    "pandas>=2.1.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "email-validator>=2.0.0",
    "schedule>=1.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
